# lock_manager: polish waiter manager