--- a/src/server/lock_manager/mod.rs
+++ b/src/server/lock_manager/mod.rs
@@ -259,21 +259,20 @@ impl LockManagerTrait for LockManager {
 
 #[cfg(test)]
 mod tests {
+    use self::metrics::*;
+    use self::waiter_manager::tests::*;
     use super::*;
     use crate::raftstore::coprocessor::Config as CopConfig;
     use crate::server::resolve::Callback;
-    use crate::storage::{
-        mvcc::Error as MvccError, mvcc::ErrorInner as MvccErrorInner, txn::Error as TxnError,
-        Error as StorageError, Result as StorageResult,
-    };
-    use kvproto::kvrpcpb::LockInfo;
-    use kvproto::metapb::{Peer, Region};
-    use metrics::*;
-    use raft::StateRole;
+    use tikv_util::security::SecurityConfig;
+
     use std::sync::mpsc;
     use std::thread;
     use std::time::Duration;
-    use tikv_util::security::SecurityConfig;
+
+    use futures::Future;
+    use kvproto::metapb::{Peer, Region};
+    use raft::StateRole;
 
     struct MockPdClient;
 
@@ -319,113 +318,99 @@ mod tests {
     fn test_single_lock_manager() {
         let lock_mgr = start_lock_manager();
 
-        let (tx, rx) = mpsc::channel();
-        let storage_callback = |tx: mpsc::Sender<_>| {
-            StorageCallback::Boolean(Box::new(move |result| {
-                tx.send(result).unwrap();
-            }))
-        };
-        let recv = |rx: &mpsc::Receiver<StorageResult<()>>| {
-            rx.recv_timeout(Duration::from_millis(500))
-                .unwrap()
-                .unwrap_err();
-        };
-
-        // Normal wake up
+        // Timeout
         assert_eq!(lock_mgr.has_waiter(), false);
-        lock_mgr.wait_for(
+        let (waiter, lock_info, f) = new_test_waiter(10.into(), 20.into(), 20);
+        lock_mgr.wait_for(waiter.start_ts, waiter.cb, waiter.pr, waiter.lock, true, 0);
+        assert_elapsed(
+            || expect_key_is_locked(f.wait().unwrap().unwrap().pop().unwrap(), lock_info),
+            3000,
+            3100,
+        );
+        assert_eq!(lock_mgr.has_waiter(), false);
+
+        // Wake up
+        let (waiter_ts, lock) = (
             10.into(),
-            storage_callback(tx.clone()),
-            ProcessResult::Res,
             Lock {
                 ts: 20.into(),
                 hash: 20,
             },
-            true,
-            0,
         );
+        let (waiter, lock_info, f) = new_test_waiter(waiter_ts, lock.ts, lock.hash);
+        lock_mgr.wait_for(waiter.start_ts, waiter.cb, waiter.pr, waiter.lock, true, 0);
         assert_eq!(lock_mgr.has_waiter(), true);
-        lock_mgr.wake_up(20.into(), Some(vec![20]), 20.into(), false);
-        recv(&rx);
+        lock_mgr.wake_up(lock.ts, Some(vec![lock.hash]), 30.into(), false);
+        assert_elapsed(
+            || expect_write_conflict(f.wait().unwrap(), waiter_ts, lock_info, 30.into()),
+            0,
+            100,
+        );
         assert_eq!(lock_mgr.has_waiter(), false);
 
-        // Normal deadlock
+        // Deadlock
+        let (waiter1, lock_info1, f1) = new_test_waiter(10.into(), 20.into(), 20);
         lock_mgr.wait_for(
-            30.into(),
-            storage_callback(tx.clone()),
-            ProcessResult::Res,
-            Lock {
-                ts: 40.into(),
-                hash: 40,
-            },
+            waiter1.start_ts,
+            waiter1.cb,
+            waiter1.pr,
+            waiter1.lock,
             false,
             0,
         );
+        let (waiter2, lock_info2, f2) = new_test_waiter(20.into(), 10.into(), 10);
         lock_mgr.wait_for(
-            40.into(),
-            storage_callback(tx.clone()),
-            ProcessResult::MultiRes {
-                results: vec![Err(StorageError::from(TxnError::from(MvccError::from(
-                    MvccErrorInner::KeyIsLocked(LockInfo::default()),
-                ))))],
-            },
-            Lock {
-                ts: 30.into(),
-                hash: 30,
-            },
+            waiter2.start_ts,
+            waiter2.cb,
+            waiter2.pr,
+            waiter2.lock,
             false,
             0,
         );
-        recv(&rx);
-        lock_mgr.wake_up(40.into(), Some(vec![40]), 40.into(), true);
-        recv(&rx);
+        assert_elapsed(
+            || expect_deadlock(f2.wait().unwrap(), 20.into(), lock_info2, 20),
+            0,
+            100,
+        );
+        // Waiter2 releases its lock.
+        lock_mgr.wake_up(20.into(), Some(vec![20]), 20.into(), true);
+        assert_elapsed(
+            || expect_write_conflict(f1.wait().unwrap(), 10.into(), lock_info1, 20.into()),
+            0,
+            100,
+        );
         assert_eq!(lock_mgr.has_waiter(), false);
 
-        // If it's the first lock, no detect
-        lock_mgr.wait_for(
-            50.into(),
-            storage_callback(tx.clone()),
-            ProcessResult::Res,
-            Lock {
-                ts: 60.into(),
-                hash: 60,
-            },
-            true,
-            0,
-        );
-        assert_eq!(lock_mgr.remove_from_detected(50.into()), false);
-        lock_mgr.wake_up(60.into(), Some(vec![60]), 60.into(), false);
-        recv(&rx);
-
-        // If it's not the first lock, detect deadlock
-        lock_mgr.wait_for(
-            50.into(),
-            storage_callback(tx.clone()),
-            ProcessResult::Res,
-            Lock {
-                ts: 60.into(),
-                hash: 60,
-            },
-            false,
-            0,
-        );
-        assert_eq!(lock_mgr.remove_from_detected(50.into()), true);
-        lock_mgr.wake_up(60.into(), Some(vec![60]), 60.into(), false);
-        recv(&rx);
+        // If it's the first lock, no detect.
+        // If it's not, detect deadlock.
+        for is_first_lock in &[true, false] {
+            let (waiter, _, f) = new_test_waiter(30.into(), 40.into(), 40);
+            lock_mgr.wait_for(
+                waiter.start_ts,
+                waiter.cb,
+                waiter.pr,
+                waiter.lock,
+                *is_first_lock,
+                0,
+            );
+            assert_eq!(lock_mgr.remove_from_detected(30.into()), !is_first_lock);
+            lock_mgr.wake_up(40.into(), Some(vec![40]), 40.into(), false);
+            f.wait().unwrap().unwrap_err();
+        }
 
         // If key_hashes is none, no wake up
         let prev_wake_up = TASK_COUNTER_VEC.wake_up.get();
-        lock_mgr.wake_up(70.into(), None, 70.into(), false);
+        lock_mgr.wake_up(10.into(), None, 10.into(), false);
         assert_eq!(TASK_COUNTER_VEC.wake_up.get(), prev_wake_up);
 
         // If it's non-pessimistic-txn, no clean up
         let prev_clean_up = TASK_COUNTER_VEC.clean_up.get();
-        lock_mgr.wake_up(80.into(), None, 80.into(), false);
+        lock_mgr.wake_up(10.into(), None, 10.into(), false);
         assert_eq!(TASK_COUNTER_VEC.clean_up.get(), prev_clean_up);
 
         // If the txn doesn't wait for locks, no clean up
         let prev_clean_up = TASK_COUNTER_VEC.clean_up.get();
-        lock_mgr.wake_up(80.into(), None, 80.into(), true);
+        lock_mgr.wake_up(10.into(), None, 10.into(), true);
         assert_eq!(TASK_COUNTER_VEC.clean_up.get(), prev_clean_up);
     }