--- a/src/server/lock_manager/waiter_manager.rs
+++ b/src/server/lock_manager/waiter_manager.rs
@@ -264,6 +264,8 @@ impl Waiter {
     }
 }
 
+// NOTE: Now we assume `Waiters` is not very long.
+// Maybe needs to use `BinaryHeap` or `VecDeque` instead.
 type Waiters = Vec<Waiter>;
 
 struct WaitTable {