<!--
Thank you for contributing to TiKV!

If you haven't already, please read TiKV's [CONTRIBUTING](https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md) document.

If you're unsure about anything, just ask; somebody should be along to answer within a day or two.
-->

###  What have you changed?
* If more than one waiters are waiting for the same lock, only the oldest one will be waked up immediately when the lock is released. Others will be waked up after `wake_up_delay_duration` to reduce contention and make the oldest one more likely acquires the lock. If the lock is released again, the second oldest will be waked up immediately, and so on.
* Each waiter has a maximum lifetime -- `wait_for_lock_timeout` to prevent starvation.
* Use lock hash as the key of WaitTable.
* Some refactor.

###  What is the type of the changes?
- Improvement (a change which is an improvement to an existing feature)

###  How is the PR tested?
- Unit test
- Manual test (add detailed scripts or steps below)


###  Does this PR affect documentation (docs) or should it be mentioned in the release notes?
No

###  Does this PR affect `tidb-ansible`?
No

###  Benchmark result if necessary (optional)
Sysbench:
``` sql
    con:query("CREATE TABLE wm(k INT PRIMARY KEY, v INT)")
    con:query("UPDATE wm SET v = v + 1 WHERE k = 1")
```

Master benchmark:

| threads | tps | .95 lat(ms) | max lat(ms) |
| -------- | ---- | ---- | ---- |
| 20 | 728 | 78.60 | - |
| 128 | 456 | 612.21 | 798.75 |
| 512 | 253 | 3773 | 4323 |
| 1000 | 193 | 9624 | 9933 |

This PR benchmark:

| threads | tps | .95 lat(ms) | max lat(ms) |
| -------- | ---- | ---- | ---- |
| 1 | 943 | 1.37 | 15.44 |
| 2 | 1079 | 2.26 | 15.24 |
| 20 | 1046 | 23.10 | 38.22 |
| 128 | 1029 | 150.29 | 186.97 |
| 512 | 980 | 580.02 | 683.31 |
| 1000 | 948 | 1191 | 1295 |